requirements.txt
Python 3.7

detectron2 == 0.6
pytorch == 1.8
torchvision == 0.9.0
opencv == 5.4.1.18 +
flask == 2.2.3 +
apscheduler == 3.10.1 +

numpy
json
hashlib